+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(albersdown)
+
+test_check("albersdown")